chore(types): refactor into smaller files and remove shorthands
