+++ packages/paste-types/src/index.ts
@@ -1,2 +1,8 @@
 export {SendGridThemeShape, ThemeShape} from '@twilio-paste/theme-tokens';
-export * from './tokenProps';
+
+export * from './tokens/backgroundColor';
+export * from './tokens/border';
+export * from './tokens/boxShadow';
+export * from './tokens/font';
+export * from './tokens/sizing';
+export * from './tokens/spacing';
