+++ packages/paste-types/src/tokens/backgroundColor.ts
@@ -1,0 +1,5 @@
+import {ThemeShape} from '@twilio-paste/theme-tokens';
+
+export interface BackgroundColorProps {
+  backgroundColor?: keyof ThemeShape['backgroundColors'];
+}
