+++ packages/paste-types/src/tokens/boxShadow.ts
@@ -1,0 +1,5 @@
+import {ThemeShape} from '@twilio-paste/theme-tokens';
+
+export interface BoxShadowProps {
+  boxShadow?: keyof ThemeShape['shadows'];
+}
